pdf-parser-api/requirements.txt
fastapi
uvicorn
celery
redis
pypdf
pdf2image
pytesseract
langchain-openai
pinecone
google-cloud-firestore
requests
python-dotenv
